Change socket-stats-interval-sec from double to u64; remove env var.

+++ collector/kernel/bpf_handler.cc
@@ -62,7 +62,7 @@ void BPFHandler::load_buffered_poller(
     u64 boot_time_adjustment,
     CurlEngine &curl_engine,
     NicPoller &nic_poller,
-    double socket_stats_interval_sec,
+    u64 socket_stats_interval_sec,
     CgroupHandler::CgroupSettings const &cgroup_settings,
     ProcessHandler::CpuMemIoSettings const *cpu_mem_io_settings,
     KernelCollectorRestarter &kernel_collector_restarter)
